pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "muse-protocol"
version = "0.1.0"
description = "Industrial baseline for multi-agent content authoring with i18n"
authors = [{name = "Muse Protocol Team"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "clickhouse-driver>=0.2",
    "datadog-api-client>=2.0",
    "deepl>=1.0",
    "pyyaml>=6.0",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
bb = "apps.cli:main"

[project.optional-dependencies]
dev = [
    "pre-commit>=3.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pytest-cov>=4.0",
    "bandit>=1.7",
    "safety>=2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "agents*", "integrations*", "schemas*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "clickhouse_driver.*",
    "datadog_api_client.*",
    "deepl.*",
]
ignore_missing_imports = true

